feat: Pokemon fetcher and shuffler
