+++ src/main.jsx
@@ -1,6 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
+import './reset.css'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
